Implement segmentation iterators for N-Gram search
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/pocketsphinx@8058 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/libpocketsphinx/pocketsphinx.c
@@ -368,7 +368,7 @@ pocketsphinx_seg_frames(ps_seg_t *seg, int *out_sf, int *out_ef)
 }
 
 void
-pocketsphinx_seg_prob(ps_seg_t *seg, float32 *out_pprob)
+pocketsphinx_seg_prob(ps_seg_t *seg, int32 *out_pprob)
 {
     *out_pprob = seg->prob;
 }
